convRDF: refactor main controller interface
