+++ src/main/java/org/g_node/Controller.java
@@ -8,26 +8,26 @@
  * LICENSE file in the root of the Project.
  */
 
-package org.g_node.crawler;
+package org.g_node;
 
 import java.util.Set;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Options;
 
 /**
- * Interface for command classes of individual crawlers.
+ * Interface for command classes of individual crawlers and converters.
  *
  * @author Michael Sonntag (sonntag@bio.lmu.de)
  */
 public interface Controller {
     /**
      * Method returning available commandline options of the crawler
-     * corresponding to the command class.
-     * @param regCrawlers Set of shorthand strings corresponding to
-     *  registered crawlers.
+     * or converter corresponding to the command class.
+     * @param regTools Set of shorthand strings corresponding to
+     *  registered crawlers or converters.
      * @return Constructed set of commandline options.
      */
-    Options options(Set<String> regCrawlers);
+    Options options(Set<String> regTools);
 
     /**
      * Method running the crawler corresponding to the command class.
